Numerical conversions

+++ purec/Compartment.cuh
@@ -44,9 +44,8 @@ extern __device__ compartment_simul_fxn_t Compartment_cuda_compartment_fxn_t;
  */
 extern __device__ void cuda_simul_fxn(void* _self,
                                       void** network,
- 
                                       const double global_time,
-                                      const uint64_t curr_step);
+                                      const uint_fast32_t curr_step);
 
 #endif /* CUDA */
