+++ purec/HHSpikeGABAAMechanism.cuh
@@ -40,7 +40,7 @@ extern __device__ double HHSpikeGABAAMechanism_cuda_mech_fun(void* _self,
                                                               void* pre_comp,
                                                               void* post_comp,
                                                               const double global_time,
-                                                              const uint64_t curr_step);
+                                                              const uint_fast32_t curr_step);
 
 #endif /* CUDA */
 #endif /* HHNSPIKEGABAAMECHANISM_CUH */
